requirements.txt
streamlit
python-dotenv
pydub
